requirements.txt
python-dotenv
pandas
numpy
pymongo
scikit-learn
matplotlib
seaborn
flask
xgboost
pymongo[srv]==3.12
certifi

# -e .